Add screen.py, move some functions to it
This file should eventually have all uses of (x)print and
sys.stdout.write, so that it can redirect that to the UI, when multiple
are supported
